# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-Tenant SaaS POS System API — Phase 1 (Foundation)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/sales-summary:
    get:
      operationId: getSalesSummaryReport
      tags: [reports]
      summary: Sales summary report
      parameters:
        - name: fromDate
          in: query
          schema: { type: string, format: date }
        - name: toDate
          in: query
          schema: { type: string, format: date }
        - name: customerId
          in: query
          schema: { type: string, format: uuid }
        - name: paymentMethod
          in: query
          schema: { type: string, enum: [CASH, CARD, INSTAPAY, WALLET, CREDIT] }
      responses:
        "200":
          description: Sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesSummaryReport"
  /reports/purchases-summary:
    get:
      operationId: getPurchasesSummaryReport
      tags: [reports]
      summary: Purchases summary report
      parameters:
        - name: fromDate
          in: query
          schema: { type: string, format: date }
        - name: toDate
          in: query
          schema: { type: string, format: date }
        - name: supplierId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Purchases summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchasesSummaryReport"
  /reports/inventory-stock:
    get:
      operationId: getInventoryStockReport
      tags: [reports]
      summary: Inventory stock valuation report
      parameters:
        - name: warehouseId
          in: query
          schema: { type: string, format: uuid }
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: brandId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Inventory stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryStockReport"
  /reports/low-stock:
    get:
      operationId: getLowStockReport
      tags: [reports]
      summary: Low stock alert report
      parameters:
        - name: warehouseId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Low stock items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LowStockReport"
  /reports/profit-loss:
    get:
      operationId: getProfitLossReport
      tags: [reports]
      summary: Profit & loss statement
      parameters:
        - name: fromDate
          in: query
          schema: { type: string, format: date }
        - name: toDate
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: P&L
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitLossReport"
  /reports/treasury:
    get:
      operationId: getTreasuryReport
      tags: [reports]
      summary: Treasury movements report
      parameters:
        - name: fromDate
          in: query
          schema: { type: string, format: date }
        - name: toDate
          in: query
          schema: { type: string, format: date }
        - name: accountId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Treasury movements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreasuryReport"
  /reports/expenses:
    get:
      operationId: getExpenseReport
      tags: [reports]
      summary: Expense report
      parameters:
        - name: fromDate
          in: query
          schema: { type: string, format: date }
        - name: toDate
          in: query
          schema: { type: string, format: date }
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseReport"
  /reports/top-products:
    get:
      operationId: getTopProductsReport
      tags: [reports]
      summary: Top selling products report
      parameters:
        - name: fromDate
          in: query
          schema: { type: string, format: date }
        - name: toDate
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopProductsReport"
...
components:
...
  schemas:
...
    SalesSummaryRow:
      type: object
      required: [id, invoiceNumber, date, total, paymentStatus]
      properties:
        id: { type: string, format: uuid }
        invoiceNumber: { type: string }
        date: { type: string }
        customerName: { type: ["string", "null"] }
        total: { type: string }
        paymentMethod: { type: ["string", "null"] }
        paymentStatus: { type: string }
    SalesSummaryReport:
      type: object
      required: [rows, count, total]
      properties:
        rows:
          type: array
          items: { $ref: "#/components/schemas/SalesSummaryRow" }
        count: { type: integer }
        total: { type: number }
    PurchasesSummaryRow:
      type: object
      required: [id, invoiceNumber, date, total, status]
      properties:
        id: { type: string, format: uuid }
        invoiceNumber: { type: string }
        date: { type: string }
        supplierName: { type: ["string", "null"] }
        total: { type: string }
        status: { type: string }
    PurchasesSummaryReport:
      type: object
      required: [rows, count, total]
      properties:
        rows:
          type: array
          items: { $ref: "#/components/schemas/PurchasesSummaryRow" }
        count: { type: integer }
        total: { type: number }
    InventoryStockRow:
      type: object
      required: [variantId, productName, quantity, value]
      properties:
        variantId: { type: string, format: uuid }
        productName: { type: string }
        sku: { type: ["string", "null"] }
        variantLabel: { type: ["string", "null"] }
        warehouseName: { type: ["string", "null"] }
        categoryName: { type: ["string", "null"] }
        brandName: { type: ["string", "null"] }
        quantity: { type: integer }
        reorderPoint: { type: integer }
        cost: { type: string }
        value: { type: number }
    InventoryStockReport:
      type: object
      required: [rows, totalValue, totalQuantity]
      properties:
        rows:
          type: array
          items: { $ref: "#/components/schemas/InventoryStockRow" }
        totalValue: { type: number }
        totalQuantity: { type: integer }
    LowStockRow:
      type: object
      required: [variantId, productName, quantity, reorderPoint, shortage]
      properties:
        variantId: { type: string, format: uuid }
        productName: { type: string }
        sku: { type: ["string", "null"] }
        variantLabel: { type: ["string", "null"] }
        warehouseName: { type: ["string", "null"] }
        quantity: { type: integer }
        reorderPoint: { type: integer }
        shortage: { type: integer }
    LowStockReport:
      type: object
      required: [rows, count]
      properties:
        rows:
          type: array
          items: { $ref: "#/components/schemas/LowStockRow" }
        count: { type: integer }
    ProfitLossReport:
      type: object
      required:
        [revenue, salesReturns, netRevenue, cogs, grossProfit, expenses, salaries, netProfit]
      properties:
        revenue: { type: number }
        salesReturns: { type: number }
        netRevenue: { type: number }
        cogs: { type: number }
        grossProfit: { type: number }
        expenses: { type: number }
        salaries: { type: number }
        netProfit: { type: number }
    TreasuryReportRow:
      type: object
      required: [id, date, accountName, direction, amount, balanceAfter, referenceType]
      properties:
        id: { type: string, format: uuid }
        date: { type: string }
        accountName: { type: string }
        direction: { type: string }
        amount: { type: string }
        balanceAfter: { type: string }
        referenceType: { type: string }
    TreasuryReport:
      type: object
      required: [rows, totalIn, totalOut]
      properties:
        rows:
          type: array
          items: { $ref: "#/components/schemas/TreasuryReportRow" }
        totalIn: { type: number }
        totalOut: { type: number }
    ExpenseReportRow:
      type: object
      required: [id, date, amount]
      properties:
        id: { type: string, format: uuid }
        date: { type: string }
        categoryName: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        amount: { type: string }
    ExpenseReport:
      type: object
      required: [rows, total]
      properties:
        rows:
          type: array
          items: { $ref: "#/components/schemas/ExpenseReportRow" }
        total: { type: number }
    TopProductRow:
      type: object
      required: [variantId, productName, quantitySold, revenue]
      properties:
        variantId: { type: string, format: uuid }
        productName: { type: string }
        sku: { type: ["string", "null"] }
        quantitySold: { type: integer }
        revenue: { type: number }
    TopProductsReport:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items: { $ref: "#/components/schemas/TopProductRow" }
...